+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   env: {
     browser: true,
+    es6: true,
     jest: true,
     node: true
   },