+++ packages/demo/__stories__/index.tsx
@@ -12,12 +12,34 @@ import { anchorPlugin } from '@edtr-io/plugin-anchor'
 import { blockquotePlugin } from '@edtr-io/plugin-blockquote'
 import { geogebraPlugin } from '@edtr-io/plugin-geogebra'
 import { highlightPlugin } from '@edtr-io/plugin-highlight'
+import { createImagePlugin, UploadConfig } from '@edtr-io/plugin-image'
 import { spoilerPlugin } from '@edtr-io/plugin-spoiler'
 import { textPlugin } from '@edtr-io/plugin-text'
 import { Overlay, rowsPlugin } from '@edtr-io/ui'
 import { storiesOf } from '@storybook/react'
 import * as React from 'react'
 
+interface SerloResponse {
+  files: { location: string }[]
+}
+
+const uploadConfig: UploadConfig<SerloResponse> = {
+  url: 'https://de.serlo.org/attachment/upload',
+  paramName: 'attachment[file]',
+  maxFileSize: 2 * 1024 * 1024,
+  allowedExtensions: ['gif', 'jpg', 'jpeg', 'png', 'svg'],
+  getAdditionalFields: () => {
+    return {
+      type: 'file',
+      csrf: ((window as unknown) as { csrf: string }).csrf
+    }
+  },
+  getStateFromResponse: response => {
+    return {
+      src: response.files[0].location
+    }
+  }
+}
 const counterState = StateType.number(0)
 
 const counterPlugin: StatefulPlugin<typeof counterState> = {
@@ -49,6 +71,7 @@ const plugins: Record<string, Plugin<any>> = {
   blockquote: blockquotePlugin,
   counter: counterPlugin,
   highlight: highlightPlugin,
+  image: createImagePlugin({ upload: uploadConfig }),
   rows: rowsPlugin,
   spoiler: spoilerPlugin,
   text: textPlugin,